+++ datatypes-cose/src/commonMain/kotlin/at/asitplus/crypto/datatypes/cose/io/Encodig.kt
@@ -1,5 +1,7 @@
 package at.asitplus.crypto.datatypes.cose.io
 
+import io.matthewnelson.encoding.base16.Base16
+import io.matthewnelson.encoding.base16.Base16ConfigBuilder
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.cbor.Cbor
 
@@ -12,3 +14,11 @@ val cborSerializer by lazy {
         writeDefiniteLengths = true
     }
 }
+
+
+/**
+ * Strict Base16 encoder
+ */
+val Base16Strict = Base16(config = Base16ConfigBuilder().apply {
+    strict()
+}.build())