+++ datatypes-jws/src/commonMain/kotlin/at/asitplus/crypto/datatypes/jws/JwsHeader.kt
@@ -45,10 +45,29 @@ data class JwsHeader(
     @Serializable(with = InstantLongSerializer::class)
     val expiration: Instant? = null,
     @SerialName("jwk")
-    val jsonWebKey: JsonWebKey? = null
+    val jsonWebKey: JsonWebKey? = null,
+    @SerialName("jku")
+    val jsonWebKeySetUrl: String? = null,
+    /**
+     * OID4VP: Verifier Attestation JWT, used to authenticate a Verifier, by providing a JWT signed by a trusted
+     * third party. May be parsed as a [JwsSigned], with [JsonWebToken] as the payload.
+     */
+    @SerialName("jwt")
+    val attestationJwt: String? = null,
 ) {
 
     fun serialize() = jsonSerializer.encodeToString(this)
+
+    /**
+     * Tries to compute a public key in descending order from [jsonWebKey], [keyId],
+     * or [certificateChain], and takes the first success or null.
+     */
+    val publicKey: CryptoPublicKey? by lazy {
+        jsonWebKey?.toCryptoPublicKey()?.getOrNull()
+            ?: keyId?.let { runCatching { CryptoPublicKey.fromDid(it) } }?.getOrNull()
+            ?: certificateChain?.firstNotNullOfOrNull { X509Certificate.decodeFromByteArray(it)?.publicKey }
+    }
+
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
@@ -56,9 +75,9 @@ data class JwsHeader(
 
         other as JwsHeader
 
-        if (algorithm != other.algorithm) return false
         if (keyId != other.keyId) return false
         if (type != other.type) return false
+        if (algorithm != other.algorithm) return false
         if (contentType != other.contentType) return false
         if (certificateChain != null) {
             if (other.certificateChain == null) return false
@@ -67,43 +86,27 @@ data class JwsHeader(
         if (notBefore != other.notBefore) return false
         if (issuedAt != other.issuedAt) return false
         if (expiration != other.expiration) return false
-        return jsonWebKey == other.jsonWebKey
+        if (jsonWebKey != other.jsonWebKey) return false
+        if (jsonWebKeySetUrl != other.jsonWebKeySetUrl) return false
+        if (attestationJwt != other.attestationJwt) return false
+
+        return true
     }
 
     override fun hashCode(): Int {
-        var result = algorithm.hashCode()
-        result = 31 * result + (keyId?.hashCode() ?: 0)
+        var result = keyId?.hashCode() ?: 0
         result = 31 * result + (type?.hashCode() ?: 0)
+        result = 31 * result + algorithm.hashCode()
         result = 31 * result + (contentType?.hashCode() ?: 0)
         result = 31 * result + (certificateChain?.contentDeepHashCode() ?: 0)
         result = 31 * result + (notBefore?.hashCode() ?: 0)
         result = 31 * result + (issuedAt?.hashCode() ?: 0)
         result = 31 * result + (expiration?.hashCode() ?: 0)
         result = 31 * result + (jsonWebKey?.hashCode() ?: 0)
+        result = 31 * result + (jsonWebKeySetUrl?.hashCode() ?: 0)
+        result = 31 * result + (attestationJwt?.hashCode() ?: 0)
         return result
     }
-
-    /**
-     * Tries to compute a public key in descending order from JWK, KeyID or the certificate chain
-     * and takes the first success or null
-     */
-    val publicKey: CryptoPublicKey? by lazy {
-        jsonWebKey?.toCryptoPublicKey()?.getOrNull()
-            ?: keyId?.let { runCatching { CryptoPublicKey.fromDid(it) } }?.getOrNull()
-            ?: certificateChain
-                ?.firstNotNullOfOrNull {
-                    runCatching {
-                        X509Certificate.decodeFromTlv(Asn1Element.parse(it) as Asn1Sequence).publicKey
-                    }.getOrNull() ?: runCatching {
-                        X509Certificate.decodeFromTlv(Asn1Element.parse(it.decodeX5c()) as Asn1Sequence).publicKey
-                    }.getOrNull()
-                }
-    }
-
-    private fun ByteArray.decodeX5c() = decodeToString()
-        .replace("-----BEGIN CERTIFICATE-----\n", "")
-        .replace("\n-----END CERTIFICATE-----", "")
-        .decodeToByteArray(Base64())
 
     companion object {
         fun deserialize(it: String) = kotlin.runCatching {